Remove conversion in to_int64

+++ csrc/_dd_ufunc.c
@@ -812,9 +812,9 @@ static void to_int64(void *_from, void *_to, npy_intp n,
             double c = f.hi - (double)thi;
 
             if(f.hi >= 0) {
-                to[i] = (uint64_t)(thi + (int64_t)floor(f.lo + c));
+                to[i] = thi + (int64_t)floor(f.lo + c);
             } else {
-                to[i] = (uint64_t)(thi + (int64_t)ceil(f.lo + c));
+                to[i] = thi + (int64_t)ceil(f.lo + c);
             }
         }
     }
